+++ tap_circle_ci/helpers.py
@@ -1,0 +1,17 @@
+import os
+import json
+
+def get_abs_path(path: str) -> str:
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
+
+# Load schemas from schemas folder
+def load_schemas() -> dict:
+    schemas = {}
+
+    for filename in os.listdir(get_abs_path('schemas')):
+        path = get_abs_path('schemas') + '/' + filename
+        file_raw = filename.replace('.json', '')
+        with open(path) as file:
+            schemas[file_raw] = json.load(file)
+
+    return schemas
